update: 代码细节优化, 后续陆续将本地代码优化后进行上传

+++ yunju-club-subject/yunju-club-infra/src/main/java/com/zsyj/subject/infra/basic/entity/SubjectMapping.java
@@ -16,11 +16,11 @@ public class SubjectMapping implements Serializable {
     /**
      * 主键
      */
-    private Integer id;
+    private Long id;
     /**
      * 题目id
      */
-    private Integer subjectId;
+    private Long subjectId;
     /**
      * 分类id
      */
