+++ yunju-club-subject/yunju-club-infra/src/main/java/com/zsyj/subject/infra/basic/entity/SubjectMultiple.java
@@ -14,11 +14,11 @@ import java.io.Serializable;
 @Data
 public class SubjectMultiple implements Serializable {
 
-    private Integer id;
+    private Long id;
     /**
      * 题目id
      */
-    private Integer subjectId;
+    private Long subjectId;
     /**
      * 选项类型（A B C D E）
      */
