+++ yunju-club-subject/yunju-club-infra/src/main/java/com/zsyj/subject/infra/basic/mapper/SubjectBriefDao.java
@@ -18,7 +18,7 @@ public interface SubjectBriefDao {
      * @param id 主键
      * @return 实例对象
      */
-    SubjectBrief queryById(Integer id);
+    SubjectBrief queryById(Long id);
 
 
 
@@ -69,7 +69,7 @@ public interface SubjectBriefDao {
      * @param id 主键
      * @return 影响行数
      */
-    int deleteById(Integer id);
+    int deleteById(Long id);
 
 }
