+++ yunju-club-subject/yunju-club-infra/src/main/java/com/zsyj/subject/infra/basic/mapper/SubjectInfoDao.java
@@ -20,7 +20,7 @@ public interface SubjectInfoDao {
      * @param id 主键
      * @return 实例对象
      */
-    SubjectInfo queryById(Integer id);
+    SubjectInfo queryById(Long id);
 
 
     /**
