+++ yunju-club-subject/yunju-club-infra/src/main/java/com/zsyj/subject/infra/basic/service/SubjectInfoService.java
@@ -19,7 +19,7 @@ public interface SubjectInfoService {
      * @param id 主键
      * @return 实例对象
      */
-    SubjectInfo queryById(Integer id);
+    SubjectInfo queryById(Long id);
